requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
groq==0.9.0
openai-whisper
gtts==2.5.1
python-dotenv==1.0.1